pyproject.toml
[tool.poetry]
name = "lemmylib"
version = "0.0.5"
description = ""
authors = ["Rooki <rooki@rooki.xyz>"]
readme = "README.md"
packages = [{include = "lemmylib"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
trylib = "lemmylib.trylib:__main__"

######### BUILD TOOLING BELOW, IMPORTANT STUFF ABOVE ######################
###########################################################################

# Test Tooling below
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = "test/"
log_cli = true
log_cli_level = "DEBUG"
#log_cli_format = "%(message)s"
log_cli_format = "[%(asctime)s] [%(levelname)8s] --- %(message)s (%(filename)s:%(funcName)s():%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.flake8]
#ignore = ['E231', 'E241']
exclude =['./tests']
max-line-length = 120
count = false
statistics = true
diff = true
format = "pylint"


[tool.black]
line-length = 120

[tool.isort]
multi_line_output = 3
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"
line_length = 120
profile = "black"
add_imports = "from __future__ import annotations"

[tool.tartufo]
repo-path = "."
regex = true
entropy = true
exclude-path-patterns = [
    {path-pattern = 'poetry\.lock'},
    {path-pattern = 'pyproject\.toml'},
    # To not have to escape `\` in regexes, use single quoted
    # TOML 'literal strings'
    {path-pattern = 'docs/source/(.*)\.rst'},
    {path-pattern = '.env\.example'},
    {path-pattern = '.env\_example'},

]
exclude-entropy-patterns = [
    {path-pattern = '\.github/workflows/.*\.yml', pattern = 'uses: .*@[a-zA-Z0-9]{40}', reason = 'GitHub Actions'},
    {path-pattern = 'poetry\.lock', pattern = '.'},
    {path-pattern = 'Pipfile\.lock', pattern = '.'},
    {path-pattern = 'README\.md', pattern = '.'},
    {path-pattern = 'mapping\.cfg', pattern = '.'}, # Git secret
    {path-pattern = '.env\.example',pattern = '.'},
]

[tool.poetry.group.test.dependencies]
setuptools = "*"
pytest = "*"
pytest-cov = "*"
pytest-pycharm = "*"
python-dotenv = "*"
pre-commit = "*"
vulture = "*"
isort = "*"
radon = "*"
xenon = "*"
black = "*"
typing_extensions ="*"
#typeguard ="*" # Typing
#mypy = "*" # Typing
pdoc3 = "*"
pylint = "*"
bandit = "*"
whispers = "*"
flake8 = "*"
Flake8-pyproject="*"
tartufo="*"
